+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - Move `ML99_TERMS` from `util.h` to `lang.h`.
  - Return a list of tuples from `ML99_listZip`, accept a list of tuples in `ML99_listUnzip`, return a tuple of lists from `ML99_listPartition`.
+ - `ML99_listEval` -> `ML99_LIST_EVAL`, `ML99_listEvalCommaSep` -> `ML99_LIST_EVAL_COMMA_SEP`
 
 ### Fixed